+++ Components/Core/Conditions/Categories/Spells.lua
@@ -35,7 +35,12 @@ local bit_band = bit.band
 local COMBATLOG_OBJECT_TYPE_PLAYER = COMBATLOG_OBJECT_TYPE_PLAYER
 
 local GetSpellCooldown = TMW.COMMON.Cooldowns.GetSpellCooldown
+local GetSpellCharges = TMW.COMMON.Cooldowns.GetSpellCharges
+local GetSpellCastCount = TMW.COMMON.Cooldowns.GetSpellCastCount
 Env.GetSpellCooldown = GetSpellCooldown
+Env.GetSpellCharges = GetSpellCharges
+Env.GetSpellCastCount = GetSpellCastCount
+
 local GetSpellName = TMW.GetSpellName
 local GetSpellInfo = TMW.GetSpellInfo
 
@@ -56,11 +61,11 @@ function Env.CooldownDuration(spell, gcdAsUnusable)
 	return 0
 end
 
-local GetSpellCharges = TMW.GetSpellCharges
 function Env.RechargeDuration(spell)
-	local charges, maxCharges, start, duration = GetSpellCharges(spell)
-	if charges and charges ~= maxCharges then
-		return (duration == 0 and 0) or (duration - (TMW.time - start))
+	local charges = GetSpellCharges(spell)
+	if charges and charges.currentCharges ~= charges.maxCharges then
+		local duration = charges.cooldownDuration
+		return (duration == 0 and 0) or (duration - (TMW.time - charges.cooldownStartTime))
 	end
 	return 0
 end
@@ -184,13 +189,17 @@ if TMW.isRetail then
 		icon = "Interface\\Icons\\ability_monk_roll",
 		tcoords = CNDT.COMMON.standardtcoords,
 		Env = {
-			GetSpellCharges = TMW.GetSpellCharges,
-			GetSpellCount = C_Spell.GetSpellCastCount or _G.GetSpellCount
+			GetSpellChargesOrCount = function(spell)
+				local charges = GetSpellCharges(spell)
+				if charges then return charges.currentCharges end
+				return GetSpellCastCount(spell)
+			end,
 		},
-		funcstr = [[(GetSpellCharges(c.OwnSpells.First) or GetSpellCount(c.OwnSpells.First)) c.Operator c.Level]],
+		funcstr = [[(GetSpellChargesOrCount(c.OwnSpells.First)) c.Operator c.Level]],
 		events = function(ConditionObject, c)
 			return
-				ConditionObject:GenerateNormalEventString("TMW_SPELL_UPDATE_CHARGES")
+				ConditionObject:GenerateNormalEventString("TMW_SPELL_UPDATE_CHARGES"),
+				ConditionObject:GenerateNormalEventString("TMW_SPELL_UPDATE_COUNT")
 		end,	
 	})
 	ConditionCategory:RegisterCondition(2.6, "SPELLCHARGETIME", {
@@ -213,17 +222,14 @@ if TMW.isRetail then
 		end),
 		icon = "Interface\\Icons\\ability_warlock_handofguldan",
 		tcoords = CNDT.COMMON.standardtcoords,
-		Env = {
-			GetSpellCharges = TMW.GetSpellCharges,
-		},
 		funcstr = [[RechargeDuration(c.OwnSpells.First) c.Operator c.Level]],
 		events = function(ConditionObject, c)
 			return
 				ConditionObject:GenerateNormalEventString("TMW_SPELL_UPDATE_CHARGES")
 		end,
 		anticipate = [[
-			local _, _, start, duration = GetSpellCharges(c.OwnSpells.First)
-			local VALUE = duration and start + (duration - c.Level) or huge
+			local data = GetSpellCharges(c.OwnSpells.First)
+			local VALUE = data and data.cooldownDuration and data.cooldownStartTime + (data.cooldownDuration - c.Level) or huge
 		]],
 	})
 end