+++ Options/CHANGELOG.lua
@@ -4,7 +4,8 @@ TMW.CHANGELOG_LASTVER="10.0.0"
 
 TMW.CHANGELOG = [==[
 ## v11.0.7
-* Fix: #2217 error in item cooldown conditions
+* Fix: #2215 - Spell Charges condition not updating for countable spells without true charges.
+* Fix: #2217 - Error in item cooldown conditions
 
 ## v11.0.6
 * #2190: Added options to Buff/Debuff icons to source stack count from tooltip numbers.